+++ HISTORY.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Add support for repeated parameters (#1).
 * *Backwards-incompatible*: All `parse_*` methods take `arg` as their fourth argument.
+* Add ``error_handler`` param to ``Parser``.
 
 0.2.0 (2014-02-26)
 ++++++++++++++++++